test: convert map unit and visual tests to TypeScript
